[Android] 가사를 통한 시간 접근구현

+++ VisibleVoice/app/src/main/java/com/example/visiblevoice/Data/Lyric.java
@@ -1,12 +1,12 @@
 package com.example.visiblevoice.Data;
 
 public class Lyric {
-    private int startTime; // 시작 밀리세컨드
+    private float startTime; // 시작 밀리세컨드
     //private int finishTime; // 끝나는 밀리세컨드
     private String text; // 해당 구간 가사
 
     public Lyric(){ }
-    public Lyric(int startTime, String text){
+    public Lyric(float startTime, String text){
         this.startTime=startTime;
         this.text=text;
     }
@@ -20,6 +20,6 @@ public class Lyric {
     public void setText(String text) { this.text = text; }
     //public int getFinishTime() { return finishTime; }
     //public void setFinishTime(int finishTime) { this.finishTime = finishTime; }
-    public int getStartTime() { return startTime; }
-    public void setStartTime(int startTime) { this.startTime = startTime; }
+    public float getStartTime() { return startTime; }
+    public void setStartTime(float startTime) { this.startTime = startTime; }
 }
